format cleanup
