+++ card/lib/card/format.rb
@@ -6,6 +6,11 @@ class Card
     include Nest
     include Permission
     include Render
+    include Names
+    include Content
+    include Error
+    extend Nest::ClassMethods
+    extend Registration
 
     DEPRECATED_VIEWS = { view: :open, card: :open, line: :closed,
                          bare: :core, naked: :core }.freeze
@@ -13,6 +18,7 @@ class Card
     # FIXME: should be set in views
 
     cattr_accessor :ajax_call, :registered
+    self.registered = []
     [:perms, :denial, :closed, :error_code,
      :view_tags, :aliases].each do |accessor_name|
       cattr_accessor accessor_name
@@ -22,67 +28,9 @@ class Card
     attr_reader :card, :root, :parent, :main_opts
     attr_accessor :form, :error_status, :nest_opts
 
-    class << self
-      @@registered = []
-
-      def register format
-        @@registered << format.to_s
-      end
-
-      def format_class_name format
-        format = format.to_s
-        format = "" if format == "base"
-        format = @@aliases[format] if @@aliases[format]
-        "#{format.camelize}Format"
-      end
-
-      def format_sym format
-        return format if format.is_a? Symbol
-        match = format.to_s.match(/::(?<format>[^:]+)Format/)
-        match ? match[:format] : :base
-      end
-
-      def extract_class_vars view, opts
-        return unless opts.present?
-        [:perms, :error_code, :denial, :closed].each do |varname|
-          class_var = send varname
-          class_var[view] = opts.delete(varname) if opts[varname]
-        end
-        extract_view_tags view, opts
-      end
-
-      def extract_view_tags view, opts
-        tags = opts.delete :tags
-        return unless tags
-        Array.wrap(tags).each do |tag|
-          view_tags[view] ||= {}
-          view_tags[view][tag] = true
-        end
-      end
-
-      def new card, opts={}
-        if self != Format
-          super
-        else
-          format = opts[:format] || :html
-          klass = Card.const_get format_class_name(format)
-          self == klass ? super : klass.new(card, opts)
-        end
-      end
-
-      def tagged view, tag
-        return unless view && tag && (view_tags = @@view_tags[view.to_sym])
-        view_tags[tag.to_sym]
-      end
-
-      def format_ancestry
-        ancestry = [self]
-        ancestry += superclass.format_ancestry unless self == Card::Format
-        ancestry
-      end
-
-      def max_depth
-        Card.config.max_depth
+    def page view, slot_opts
+      @card.run_callbacks :show_page do
+        show view, slot_opts
       end
     end
 
@@ -103,22 +51,9 @@ class Card
       @root ||= self
       @depth ||= 0
 
-      @context_names = get_context_names
+      @context_names = initialize_context_names
       include_set_format_modules
       self
-    end
-
-    def get_context_names
-      case
-      when @context_names
-        part_keys = @card.cardname.part_names.map(&:key)
-        @context_names.reject { |n| !part_keys.include? n.key }
-      when params[:slot]
-        context_name_list = params[:slot][:name_context].to_s
-        context_name_list.split(",").map(&:to_name)
-      else
-        []
-      end
     end
 
     def include_set_format_modules
@@ -139,22 +74,6 @@ class Card
 
     def session
       Env.session
-    end
-
-    def showname title=nil
-      if title
-        title.to_name.to_absolute_name(card.cardname).to_show(*@context_names)
-      else
-        @showname ||= card.cardname.to_show(*@context_names)
-      end
-    end
-
-    def with_name_context name
-      old_context = @context_names
-      add_name_context name
-      result = yield
-      @context_names = old_context
-      result
     end
 
     def main?
@@ -180,68 +99,20 @@ class Card
 
     def method_missing method, *opts, &proc
       if method =~ /(_)?(optional_)?render(_(\w+))?/
-        view = Regexp.last_match(3) ? Regexp.last_match(4) : opts.shift
-        args = opts[0] ? opts.shift.clone : {}
-        if Regexp.last_match(2)
-          args[:optional] = true
-          args[:default_visibility] = opts.shift
-        end
-        args[:skip_permissions] = true if Regexp.last_match(1)
-        render view, args
+        render_api Regexp.last_match, opts
       else
-        proc = proc { |*a| raw yield(*a) } if proc
-        response = root.template.send method, *opts, &proc
-        response.is_a?(String) ? root.template.raw(response) : response
+        pass_method_to_template_object(method, opts, proc) { yield }
       end
     end
 
-    #
-    # ------------- Sub Format and Inclusion Processing ------------
-    #
-
-    def process_content override_content=nil, opts={}
-      content = override_content || render_raw || ""
-      content_object = get_content_object content, opts
-      content_object.process_each_chunk do |chunk_opts|
-        prepare_nest chunk_opts.merge(opts)
-      end
-      content_object.to_s
-    end
-
-    def get_content_object content, opts
-      if content.is_a? Content
-        content
-      else
-        Content.new content, self, opts.delete(:content_opts)
-      end
+    def pass_method_to_template_object method, opts, proc
+      proc = proc { |*a| raw yield(*a) } if proc
+      response = root.template.send method, *opts, &proc
+      response.is_a?(String) ? root.template.raw(response) : response
     end
 
     def tagged view, tag
       self.class.tagged view, tag
-    end
-
-    #
-    # ------------ LINKS ---------------
-    #
-
-    def format_date date, include_time=true
-      # using DateTime because Time doesn't support %e on some platforms
-      if include_time
-        # .strftime('%B %e, %Y %H:%M:%S')
-        I18n.localize(DateTime.new(date.year, date.mon, date.day,
-                                   date.hour, date.min, date.sec),
-                      format: :card_date_seconds)
-      else
-        # .strftime('%B %e, %Y')
-        I18n.localize(DateTime.new(date.year, date.mon, date.day),
-                      format: :card_date_only)
-      end
-    end
-
-    def add_name_context name=nil
-      name ||= card.name
-      @context_names += name.to_name.part_names
-      @context_names.uniq!
     end
   end
 end
